splitted classes in person.hpp, implemented my_info in person, student and mathematician. all_info in person calls person::my_info, all_info in student calls person::my_info and my_info, math_student all_info calls student::all_info and mathematician::my_info, works as expected.
