+++ week5/src/CMakeLists.txt
@@ -5,4 +5,4 @@ project (Week5)
 SET(GCC_COVERAGE_COMPILE_FLAGS "-std=c++11 -O0 -g") # Debug
 SET( CMAKE_CXX_FLAGS  "${CMAKE_CXX_FLAGS} ${GCC_COVERAGE_COMPILE_FLAGS}" )
 
-add_executable(test test.cpp)
+add_executable(test test.cpp person.hpp)
